Added new component file and moved code around
